requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.11.0=pypi_0
asn1crypto=1.4.0=pyh9f0ad1d_0
astunparse=1.6.3=pypi_0
backcall=0.2.0=py_0
blas=1.0=mkl
blessings=1.7=pypi_0
brotlipy=0.7.0=py37hb5d75c8_1001
bzip2=1.0.8=h516909a_3
ca-certificates=2021.5.30=ha878542_0
cached-property=1.5.2=pypi_0
cachetools=4.2.0=pypi_0
cairo=1.14.12=h8948797_3
certifi=2021.5.30=py37h89c1867_0
cffi=1.14.4=py37h11fe52a_0
chardet=4.0.0=py37h89c1867_1
cloudpickle=1.6.0=py_0
cryptography=2.5=py37hb7f436b_1
cudatoolkit=11.0.221=h6bb024c_0
cvxpy=1.1.7=py37h3340039_0
cvxpy-base=1.1.7=py37h9fdb41a_0
cycler=0.10.0=py37_0
cytoolz=0.11.0=py37h7b6447c_0
dask-core=2021.4.0=pyhd3eb1b0_0
dataclasses=0.8=pyhc8e2a94_1
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=py_0
dotmap=1.3.23=pypi_0
ecos=2.0.7.post1=py37ha21ca33_3
expat=2.2.10=he6710b0_2
ffmpeg=4.0=hcdf2ecd_0
flatbuffers=1.12=pypi_0
fontconfig=2.13.0=h9420a91_0
freeglut=3.2.1=h58526e2_0
freetype=2.10.4=h5ab3b9f_0
fsspec=0.9.0=pyhd3eb1b0_0
future=0.18.2=py37h89c1867_3
gast=0.4.0=pypi_0
glib=2.63.1=h5a9c865_0
google-api-core=1.26.3=pypi_0
google-api-python-client=2.2.0=pypi_0
google-auth=1.24.0=pypi_0
google-auth-httplib2=0.1.0=pypi_0
google-auth-oauthlib=0.4.4=pypi_0
google-pasta=0.2.0=pypi_0
googleapis-common-protos=1.53.0=pypi_0
gpustat=0.6.0=pypi_0
graphite2=1.3.13=h58526e2_1001
grpcio=1.34.1=pypi_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=3.1.0=pypi_0
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hc401514_3
httplib2=0.19.1=pypi_0
icu=58.2=he6710b0_3
idna=2.10=pyh9f0ad1d_0
imageio=2.9.0=pyhd3eb1b0_0
importlib-metadata=3.3.0=pypi_0
intel-openmp=2020.2=254
ipdb=0.13.4=pyhd3deb0d_0
ipython=7.19.0=py37hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jasper=2.0.14=h07fcdf6_1
jedi=0.17.2=py37h89c1867_1
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9b=h024ee3a_2
jsonpatch=1.28=pyhd3eb1b0_0
jsonpath=0.82=pypi_0
jsonpointer=2.0=py_0
keras=2.4.3=pypi_0
keras-nightly=2.5.0.dev2021032900=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.0=py37h2531618_0
lcms2=2.11=h396b838_0
libblas=3.8.0=21_mkl
libcblas=3.8.0=21_mkl
libedit=3.1.20191231=h14c3975_1
libffi=3.2.1=hf484d3e_1007
libgcc-ng=9.1.0=hdf63c60_0
libgfortran=3.0.0=1
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=he1b5a44_1001
liblapack=3.8.0=21_mkl
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.13.0.1=h8b12597_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
locket=0.2.1=py37h06a4308_1
lz4-c=1.9.2=heb0550a_3
markdown=3.3.3=pypi_0
matplotlib=3.3.2=h06a4308_0
matplotlib-base=3.3.2=py37h817c723_0
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.2.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
mpi=1.0=mpich
mpi4py=3.0.3=py37hf046da1_1
mpich=3.3.2=hc856adb_0
natsort=7.1.1=pyhd3eb1b0_0
ncurses=6.2=he6710b0_1
networkx=2.5.1=pyhd3eb1b0_0
ninja=1.10.2=py37hff7bd54_0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
nvidia-ml-py3=7.352.0=pypi_0
oauth2client=4.1.3=pypi_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py37_0
opencv=3.4.2=py37h6fd60c2_1
openssl=1.0.2u=h516909a_0
opt-einsum=3.3.0=pypi_0
ordered-set=4.0.2=pypi_0
osqp=0.6.1=py37h0da4684_2
packaging=20.9=pypi_0
pandas=1.2.1=py37ha9443f7_0
parso=0.8.1=pypi_0
partd=1.2.0=pyhd3eb1b0_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.0.1=py37he98fc37_0
pip=20.3.3=py37h06a4308_0
pixman=0.40.0=h36c2ea0_0
prompt-toolkit=3.0.8=py_0
protobuf=3.14.0=pypi_0
psutil=5.8.0=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
py-opencv=3.4.2=py37hb342d67_1
py3nvml=0.2.6=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=pyh9f0ad1d_2
pycutest=1.0=pypi_0
pydrive=1.3.1=pypi_0
pygments=2.7.3=pyhd3eb1b0_0
pyopenssl=19.0.0=py37_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h05f1152_2
pysocks=1.7.1=py37h89c1867_3
python=3.7.0=h6e4f718_3
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.7.1=py3.7_cuda11.0.221_cudnn8.0.5_0
pytorch-fid=0.2.0=pypi_0
pytorch-lightning=0.8.5=py_0
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.1.1=py37h7b6447c_2
pyyaml=5.3.1=py37hb5d75c8_1
pyzmq=20.0.0=py37h5a562af_1
qpth=0.0.15=pypi_0
qt=5.9.6=h8703b6f_2
readline=7.0=h7b6447c_5
requests=2.25.1=pyhd3deb0d_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.6=pypi_0
scikit-image=0.18.1=py37ha9443f7_0
scikit-learn=0.23.2=py37hddcf8d6_3
scipy=1.5.2=py37h0b6359f_0
scs=2.1.2=py37h26cea63_2
setproctitle=1.1.10=py37h8f50634_1003
setuptools=51.0.0=py37h06a4308_2
sip=4.19.8=py37hf484d3e_0
six=1.15.0=py37h06a4308_0
sqlite=3.33.0=h62c20be_0
tensorboard=2.5.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.7.0=pypi_0
tensorboardx=2.1=pypi_0
tensorflow=2.4.0=pypi_0
tensorflow-estimator=2.5.0=pypi_0
tensorflow-gpu=2.5.0=pypi_0
termcolor=1.1.0=py37_1
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2020.10.1=py37hdd07704_2
tk=8.6.10=hbc83047_0
toolz=0.11.1=pyhd3eb1b0_0
torchaudio=0.7.2=py37
torchfile=0.1.0=py_0
torchmeta=1.8.0=pypi_0
torchvision=0.8.2=py37_cu110
tornado=6.1=py37h27cfd23_0
tqdm=4.60.0=pyhd8ed1ab_0
traitlets=5.0.5=py_0
typing_extensions=3.7.4.3=py_0
uritemplate=3.0.1=pypi_0
urllib3=1.26.2=pypi_0
visdom=0.1.8.9=0
wcwidth=0.2.5=py_0
websocket-client=0.57.0=py37h89c1867_4
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
wrapt=1.12.1=pypi_0
xmltodict=0.12.0=pypi_0
xorg-fixesproto=5.0=h14c3975_1002
xorg-inputproto=2.3.2=h14c3975_1002
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libx11=1.6.12=h516909a_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxfixes=5.0.3=h516909a_1004
xorg-libxi=1.7.10=h516909a_0
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.5=h7b6447c_0
yacs=0.1.6=py_0
yaml=0.2.5=h516909a_0
zeromq=4.3.3=h58526e2_3
zipp=3.4.0=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
